requirements.txt
gradio~=5.30.0
requests~=2.32.3
pandas~=2.2.3
pytest~=8.3.5
python-dotenv~=1.1.0
langchain-core~=0.3.60
langgraph~=0.4.5
duckduckgo_search~=8.0.1
langchain-community~=0.3.24